Separated the endpoints into their own files for cleaner code
