+++ blockChain.py
@@ -185,90 +185,3 @@ node_identifier = str(uuid4()).replace('-','')
 
 #Instantiate the Blockchain
 blockchain = Blockchain()
-
-@app.route('/mine_block', methods=['GET'])
-def mine_block():
-    #We run the proof of work algorithm to get the next proof
-    last_block = blockchain.last_block
-    last_proof = last_block['proof']
-    proof = blockchain.proof_of_work(last_proof)
-    
-    #We must receive a reward (aka a coin) for finding the proof
-    #The sender is "0" to signify that this node has mined a new coin
-    blockchain.new_transaction(
-        sender="0",
-        recipient=node_identifier,
-        amount=1,
-    )
-    
-    #Forge the new Block by adding it to the chain
-    previous_hash = blockchain.hash(last_block)
-    block = blockchain.new_block(proof, previous_hash)
-    
-    response = {
-        'message': "New Block Forged",
-        'index': block['index'],
-        'transactions': block['transactions'],
-        'proof': block['proof'],
-        'previous_hash': block['previous_hash'],
-    }
-    return jsonify(response), 200
-
-@app.route('/add_new_transaction', methods=['POST'])
-def add_new_transaction():
-    values = request.get_json()
-    
-    #Check that the required fields are in the POST'ed data
-    required = ['sender','recipient','amount']
-    if not all(k in values for k in required):
-        return 'Missing required values', 400
-    
-    #Create a new Transaction
-    index = blockchain.new_transaction(values['sender'], values['recipient'], values['amount'])
-    
-    response = {'message': f'Transaction will be added to the Block {index}'}
-    return jsonify(response), 201
-
-@app.route('/return_chain', methods=['GET'])
-def return_chain():
-    response = {
-        'chain': blockchain.chain,
-        'length': len(blockchain.chain),
-    }
-    return jsonify(response), 200
-
-@app.route('/register_node', methods=['POST'])
-def register_nodes():    
-    values = request.get_json()
-    
-    nodes = values.get('nodes')
-    if nodes is None:
-        return "Error: Please supply a list of nodes", 400
-    
-    for node in nodes:
-        blockchain.register_node(node)
-        
-        response = {
-            'message':'New nodes have been successfully added',
-            'total_nodes':list(blockchain.nodes),
-        }
-        return jsonify(response), 201
-    
-@app.route('/resolve_node', methods=['GET'])
-def consensus():
-        replaced = blockchain.resolve_conflicts()
-        
-        if replaced:
-            response = {
-                'message': 'Our chain was replaced',
-                'chain': blockchain.chain
-            }
-        else:
-            response = {
-                'message': 'Our chain is the OG',
-                'chain': blockchain.chain
-            }
-        return jsonify(response), 200
-
-if __name__ == '__main__':
-    app.run(host='0.0.0.0', port=9000)
