+++ main.py
@@ -1,0 +1,6 @@
+from blockChain import app
+import registerNode, addNewTransaction, mineBlock, registerNode
+
+
+if __name__ == '__main__':
+    app.run(host='0.0.0.0', port=9000)
